vhost: Adding support for virtio-vhost-user transport
The virtio-vhost-user device lets guests act as vhost device backends so
that storage appliance VMs can provide virtio devices to other guests.
This way, the SPDK vhost target can be shipped into a VM. This is an
alternative to the pre-existing unix transport. More information about
this device can be found here:
https://wiki.qemu.org/Features/VirtioVhostUser

Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>