vhost: add command line option for vvu transport
Add -T option in application-specific command line options for vhost app
to choose between unix and vvu transports. Default is unix if not
specified. -S option is restructured to represent the directory for the
unix sockets to be created in case of unix transport or the PCI address
in BDF format for the virtio-vhost-user device in case of vvu transport.
Type "app/vhost/vhost -h" for more about how to use these options.

Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>