env_dpdk: add flag RTE_PCI_DRV_NEED_MAPPING in vvu driver flags
Remove the manual invocation of rte_pci_map_device() which performs the
mapping of PCI resources via vfio. Use the flag instead. When the flag
is used, the DPDK runtime performs the mapping before probing the
appropriate userspace driver for this device. The check is done by
rte_pci_probe_one_driver().

Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>