setup.sh: use vfio kernel driver in case of vfio no-IOMMU mode
In previous state, vfio driver was being used only if an IOMMU were
present. This was checked by looking for any IOMMU groups in sysfs.

In case of no-iommu driver, there are no IOMMU groups when loading the
vfio modules. The IOMMU groups are created when binding the PCI devices
to the vfio-pci driver.

In the previous state, uio_pci_generic was being used in this case
because binding is done after choosing the kernel driver.

In order to support vfio no-IOMMU mode, we check if the vfio module is
loaded with parameter "enable_unsafe_noiommu_mode". In other case, we
use uio_pci_generic driver.

Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>