bdev/virtio: add VIRTIO_F_IOMMU_PLATFORM feature bit
This commit adds the VIRTIO_F_IOMMU_PLATFORM feature bit in the
virtio-scsi bdev driver.

This feature bit had not been added so far because this driver has been
designed to control vhost-user-scsi-pci devices. The device backend of
those devices is the SPDK vhost-scsi driver. The vhost-scsi driver
currently is not supporting this feature bit. So, there was no point in
adding functionality for this feature in the virtio-scsi bdev driver
because it would not be used.

The reason I am adding this feature in the virtio-scsi bdev driver is to
enable supporting device backends other than vhost-user-scsi. For
example, QEMU user space scsi target.

With this patch we can use the virtio-scsi bdev driver to control a
virtio-scsi device terminating on QEMU scsi target and make use of a
vIOMMU for DMA translation.

However, it is not yet possible to use this driver for a virtio-scsi
device not supporting this feature and a system with a vIOMMU and vfio
enabled. The reason is that, in this case, the virtio driver would have
to pass physical DMA addresses to the device, but the vtophys map keeps
only IOVAs in case vfio is enabled. The solution would be to extend the
vtophys interface to enable finding physical addresses even when vfio is
enabled. We also need to add a conditional statement in the virtio
driver checking whether the feature is enabled and thus choosing the
right IOV base address (PA or IOVA).

Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>