vhost/rte_vhost: support RPC "remove_vhost_controller" for vvu transport
Currently, the vvu device is being attached when a vhost controller is
created. We want it to be detached when the associated vhost controller
is being removed.

This makes sense because a vvu device can support only one vhost
controller. This implies that a vvu device is tightly related to one
vhost contoller. Thus, when the controller is removed, the vvu device
can also be removed.

This commit enables detaching the vvu device when socket_cleanup() is
invoked.

socket_cleanup() is a transport-specific function that frees the
resources associated with a vhost controller's transport-specific state.
It is called in three cases:

- a new vhost controller instance fails to be created
- RPC "remove_vhost_controller" is invoked
- SIGINT signal is received

Detaching the vvu device alongside with the vhost controller's removal
has two advantages:

- the vvu device stays attached only for the time interval it is needed.
  This is an SPDK policy. Devices are being used only for the time they
  are needed.
- RPC "remove_vhost_controller" works as expected. This means that
  reusing the same vvu device to construct a new vhost controller is now
  possible.

Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>