dpdk: update submodule
Include a patch that restricts the mappings of the PCI BARs of all PCI
devices to be 2MB-aligned in the process virtual address space. This
allows the vhost library to support the virtio-vhost-user transport.

This is a workaround over a restriction of the vtophys map. The vtophys
map registers mappings only for 2MB-aligned virtual addresses. This is
not a problem in case of the AF_UNIX transport because the VM's memory
is backed by hugepages and thus the mmapped virtual addresses of the
vhost memory regions are 2MB-aligned for sure. However, in case of the
virtio-vhost-user transport, the VM's memory is exposed to the slave
guest as a PCI memory region of the virtio-vhost-user device. Mapping
this PCI MMIO region through the vfio device file does not guarantee
that the mmap address will be 2MB-aligned. As a result, FLOOR_2MB() in
spdk_vhost_session_mem_register() may return a virtual address with no
corresponding physical address. This either triggers a SEGFAULT in
vtophys_get_paddr_pagemap() when the page is touched in case of vfio
no-IOMMU mode or causes the VFIO_IOMMU_MAP_DMA ioctl to fail in case of
vfio type1.

This patch overrides this problem by mapping all PCI resources of all
PCI devices to 2MB-aligned virtual addresses.

A more suitable solution would be to extend the vtophys API to support
registering memory mappings for non 2MB-aligned virtual addresses.

Change-Id: Ie4b503f5a5e74f95076b97c2e98ac58772bd33cd
Signed-off-by: Nikos Dragazis <ndragazis@outlook.com.gr>